Fix GitHub Actions build errors and compiler warnings
Critical fixes:
- Add missing 'static' target to codegens/web/Makefile
- Fix STB image linking conflict between raylib and SDL3 backend
  - Add KRYON_USE_EXTERNAL_STB guard in stb_image_impl.c
  - Pass -DKRYON_USE_EXTERNAL_STB from SDL3 Makefile
- Fix wrong enum type names in ir_capability.c (KRYON_COMPONENT_* → IR_COMPONENT_*)

IR library warnings fixed (30+ files):
- Unused variables: ir_string_builder.c, ir_markdown.c, ir_validation.c, ir_layout.c, ir_stylesheet.c, markdown_to_ir.c, kry_to_ir.c, css_generator_core.c
- Unused functions: ir_builder.c, ir_tabgroup.c, ir_instance.c, ir_format_detect.c, ir_builtin_registry.c, ir_stylesheet.c, ir_state_manager.c, ir_expression_compiler.c, foreach_expand.c, markdown_parser.c, html_parser.c, kry_to_ir.c, strategy_row.c
- Unused parameters: ir_stylesheet.c, ir_state_manager.c, kry_codegen.c, markdown_parser.c
- Sign comparisons: ir_tabgroup.c, ir_buffer.c, ir_layout.c, ir_executor.c, ir_stylesheet.c, kry_to_ir.c, strategy_container.c, ir_web_renderer.c
- Missing field initializer: ir_component_dropdown.c
- Incomplete switch: ir_builtin_registry.c
- Removed orphaned forward declarations: ir_instance.c, ir_json_deserialize.c

CLI warnings fixed:
- Check return values of system() calls in dev_server.c, plugin_installer.c, build.c, luajit_link.c, luajit_compile.c, cmd_plugin.c, cmd_run.c
- Check return values of fread() in cmd_compile.c, build.c
- Check return value of fgets() in cmd_run.c
- Fix strncpy warnings in plugin_discovery.c, cmd_run.c, luajit_link.c
- Add format truncation checks for snprintf calls
- Add __attribute__((unused)) to free_loaded_plugin_tracking in config.c

Web codegen warnings fixed:
- Add __attribute__((unused)) to base_len in css_generator_core.c
- Fix sign comparison in ir_web_renderer.c
- Check write() return value in html_generator.c
- Replace strncpy with memcpy in lua_bundler.c

+++ codegens/web/ir_web_renderer.c
@@ -318,7 +318,7 @@ static void collect_assets(IRComponent* component, char** assets, int* count, in
     }
 
     // Recurse into children
-    for (int i = 0; i < component->child_count; i++) {
+    for (int i = 0; i < (int)component->child_count; i++) {
         collect_assets(component->children[i], assets, count, max);
     }
 }
